[Keras] clean up, move some functions into base_model
