+++ python/flexflow/keras/models/base_model.py
@@ -1,6 +1,7 @@
 import flexflow.core as ff
 
 from .input_layer import Tensor
+from flexflow.keras.layers import Conv2D, MaxPooling2D, Flatten, Dense, Activation, Concatenate
 from flexflow.keras.optimizers import SGD, Adam 
 
 from PIL import Image
@@ -26,6 +27,14 @@ class BaseModel(object):
     self.label_dataloader = 0
     self.label_dataloader_dim = 0
     
+  @property
+  def input(self):
+    return self.input_tensors
+  
+  @property
+  def output(self):
+    return self.output_tensor
+    
   def get_layer(self, layer_id):
     return self._layers[layer_id]
     
@@ -35,6 +44,15 @@ class BaseModel(object):
     
   def _create_label_tensor(self):
     self.label_tensor = Tensor(self.ffmodel, batch_shape=[self.ffconfig.get_batch_size(), 1], name="", dtype="int32")
+    
+  def _create_input_and_label_tensors(self):
+    idx = 0
+    for input_tensor in self.input_tensors:
+      input_tensor.set_batch_size(self.ffconfig.get_batch_size())
+      self._create_input_tensor(idx)
+      idx += 1
+
+    self._create_label_tensor()
     
   def _verify_tensors(self, input_arrays, label_array):
     assert len(input_arrays) == len(self.input_tensors), "check len of input tensors"
@@ -133,8 +151,8 @@ class BaseModel(object):
           self.ffconfig.begin_trace(111)
         self.ffmodel.forward()
         # for layer_id in self._layers:
-        #  layer = self._layers[layer_id]
-        #  layer.ffhandle.forward(self.ffmodel)
+        #   layer = self._layers[layer_id]
+        #   layer.ffhandle.forward(self.ffmodel)
         self.ffmodel.zero_gradients()
         self.ffmodel.backward()
         self.ffmodel.update()
@@ -157,6 +175,89 @@ class BaseModel(object):
     print(label_array.shape)
     print(label_array)
     self.label_tensor.ffhandle.inline_unmap(self.ffconfig)
+    
+  def _create_flexflow_layers_v2(self):
+    for layer_id in self._layers:
+      layer = self._layers[layer_id]
+
+      if (isinstance(layer, Conv2D) == True):
+        layer.ffhandle = self.ffmodel.conv2d_v2(layer.name, layer.in_channels, layer.out_channels, layer.kernel_size[0], layer.kernel_size[1], layer.stride[0], layer.stride[1], layer.padding[0], layer.padding[1], layer.activation, layer.use_bias)
+      elif (isinstance(layer, MaxPooling2D) == True):
+        layer.ffhandle = self.ffmodel.pool2d_v2(layer.name, layer.kernel_size[1], layer.kernel_size[0], layer.stride[0], layer.stride[1], layer.padding[0], layer.padding[1])
+      elif (isinstance(layer, Flatten) == True):
+        layer.ffhandle = self.ffmodel.flat_v2(layer.name)
+      elif (isinstance(layer, Dense) == True):
+        layer.ffhandle = self.ffmodel.dense_v2(layer.name, layer.in_channels, layer.out_channels, layer.activation)
+      elif (isinstance(layer, Activation) == True):
+        print("add softmax")
+      elif (isinstance(layer, Concatenate) == True):
+        print("add concatenate")
+      else:
+        assert 0, "unknow layer"
+        
+  def _create_flexflow_layers(self, verify_inout_shape=True):
+    out_t = 0
+    for layer_id in self._layers:
+      layer = self._layers[layer_id]
+
+      if (isinstance(layer, Activation) == True):
+       assert layer.layer_id == self._nb_layers-1, "softmax is not in the last layer"
+       out_t = self.ffmodel.softmax("softmax", layer.input_tensors[0].ffhandle, self.label_tensor.ffhandle)
+      elif (isinstance(layer, Concatenate) == True):
+       t_ffhandle_list = []
+       for t in layer.input_tensors:
+         t_ffhandle_list.append(t.ffhandle)
+       out_t = self.ffmodel.concat("concat", t_ffhandle_list, layer.axis)
+      elif (isinstance(layer, Conv2D) == True):
+       out_t = self.ffmodel.conv2d(layer.name, layer.input_tensors[0].ffhandle, layer.out_channels, layer.kernel_size[0], layer.kernel_size[1], layer.stride[0], layer.stride[1], layer.padding[0], layer.padding[1], layer.activation, layer.use_bias)
+      elif (isinstance(layer, MaxPooling2D) == True):
+       out_t = self.ffmodel.pool2d(layer.name, layer.input_tensors[0].ffhandle, layer.kernel_size[1], layer.kernel_size[0], layer.stride[0], layer.stride[1], layer.padding[0], layer.padding[1])
+      elif (isinstance(layer, Flatten) == True):
+       out_t = self.ffmodel.flat(layer.name, layer.input_tensors[0].ffhandle)
+      elif (isinstance(layer, Dense) == True):
+       out_t = self.ffmodel.dense(layer.name, layer.input_tensors[0].ffhandle, layer.out_channels, layer.activation)
+      else:
+       assert 0, "unknow layer"
+
+      layer.output_tensor.set_ffhandle(out_t)
+
+      assert layer.ffhandle == 0, "layer handle is inited"
+      layer.ffhandle = self.ffmodel.get_layer_by_id(layer.layer_id)
+      assert layer.ffhandle != 0, "layer handle is wrong"
+      print(layer.ffhandle)    
+
+      if (verify_inout_shape == True):
+       in_t = layer.input_tensors[0].ffhandle
+       layer.verify_inout_shape(in_t, out_t)
+       
+  def _init_inout(self, verify_inout_shape=True):
+    out_t = 0
+    for layer_id in self._layers:
+      layer = self._layers[layer_id]
+
+      if (isinstance(layer, Activation) == True):
+        assert layer_id == self._nb_layers-1, "softmax is not in the last layer"
+        out_t = self.ffmodel.softmax("softmax", layer.input_tensors[0].ffhandle, self.label_tensor.ffhandle)
+        assert layer.ffhandle == 0, "layer handle is inited"
+        layer.ffhandle = self.ffmodel.get_layer_by_id(layer_id)
+      elif (isinstance(layer, Concatenate) == True):
+        t_ffhandle_list = []
+        for t in layer.input_tensors:
+          t_ffhandle_list.append(t.ffhandle)
+        out_t = self.ffmodel.concat("concat", t_ffhandle_list, layer.axis)
+        assert layer.ffhandle == 0, "layer handle is inited"
+        layer.ffhandle = self.ffmodel.get_layer_by_id(layer_id)
+      else:
+        out_t = layer.ffhandle.init_inout(self.ffmodel, layer.input_tensors[0].ffhandle);
+      
+      layer.output_tensor.set_ffhandle(out_t)
+      assert layer.ffhandle != 0, "layer handle is wrong"
+      print(layer.ffhandle)    
+      
+      if (verify_inout_shape == True):
+        in_t = layer.input_tensors[0].ffhandle
+        layer.verify_inout_shape(in_t, out_t)
+    print("output tensor", self.output_tensor.batch_shape)
     
   def summary(self):
     model_summary = "Layer (type)\t\tOutput Shape\t\tInput Shape\tConnected to\n"
